+++ frontend/.prettierrc.json
@@ -18,7 +18,7 @@
     "^@utils/(.*)$",
     "^@custom-types",
     "^@styles/(.*)$",
-    "^@api/(.*)$",
+    "^@api",
     "^@hooks/(.*)$",
     "^@context/(.*)$",
     "^@layout/(.*)$",
