+++ frontend/src/api/getStudyDetail.ts
@@ -1,6 +1,6 @@
 import type { GetStudyDetailRequestParams, GetStudyDetailResponseData } from '@custom-types';
 
-import axiosInstance from '@api/axiosInstance';
+import { axiosInstance } from '@api';
 
 const getStudyDetail = async ({ studyId }: GetStudyDetailRequestParams) => {
   const response = await axiosInstance.get<GetStudyDetailResponseData>(`/api/studies/${studyId}`);
