+++ frontend/src/api/getStudyReviews.ts
@@ -1,6 +1,6 @@
 import type { GetReviewRequestParams, GetReviewResponseData } from '@custom-types';
 
-import axiosInstance from '@api/axiosInstance';
+import { axiosInstance } from '@api';
 
 const getStudyReviews = async ({ studyId, size }: GetReviewRequestParams) => {
   const url = size ? `/api/studies/${studyId}/reviews?size=${size}` : `/api/studies/${studyId}/reviews`;
