+++ frontend/src/api/patchReview.ts
@@ -2,7 +2,7 @@ import type { AxiosResponse } from 'axios';
 
 import type { EmptyObject, PatchReviewRequestBody, PatchReviewRequestVariables } from '@custom-types';
 
-import axiosInstance from '@api/axiosInstance';
+import { axiosInstance } from '@api';
 
 const patchReview = async ({ studyId, reviewId, content }: PatchReviewRequestVariables) => {
   const response = await axiosInstance.patch<EmptyObject, AxiosResponse<EmptyObject>, PatchReviewRequestBody>(
