+++ frontend/src/api/postJoiningStudy.ts
@@ -2,7 +2,7 @@ import type { AxiosResponse } from 'axios';
 
 import type { EmptyObject, PostJoiningStudyRequestParams } from '@custom-types';
 
-import axiosInstance from '@api/axiosInstance';
+import { axiosInstance } from '@api';
 
 const postJoiningStudy = async ({ studyId }: PostJoiningStudyRequestParams) => {
   const response = await axiosInstance.post<EmptyObject, AxiosResponse<EmptyObject>, PostJoiningStudyRequestParams>(
