+++ frontend/src/api/postNewStudy.ts
@@ -2,7 +2,7 @@ import type { AxiosResponse } from 'axios';
 
 import type { EmptyObject, PostNewStudyRequestBody } from '@custom-types';
 
-import axiosInstance from '@api/axiosInstance';
+import { axiosInstance } from '@api';
 
 const postNewStudy = async (data: PostNewStudyRequestBody) => {
   const response = await axiosInstance.post<EmptyObject, AxiosResponse<EmptyObject>, PostNewStudyRequestBody>(
