+++ frontend/src/api/postReview.ts
@@ -2,7 +2,7 @@ import type { AxiosResponse } from 'axios';
 
 import { EmptyObject, PostReviewRequestBody, PostReviewRequestVariables } from '@custom-types';
 
-import axiosInstance from '@api/axiosInstance';
+import { axiosInstance } from '@api';
 
 const postReview = async ({ studyId, content }: PostReviewRequestVariables) => {
   const response = await axiosInstance.post<EmptyObject, AxiosResponse<EmptyObject>, PostReviewRequestBody>(
