+++ frontend/src/context/login/LoginProvider.tsx
@@ -1,8 +1,7 @@
+import { noop } from '@utils';
 import { ReactNode, createContext, useState } from 'react';
 
 import { ACCESS_TOKEN_KEY } from '@constants';
-
-import { noop } from '@utils/index';
 
 type LoginProviderProps = {
   children: ReactNode;
