+++ frontend/src/layout/header/components/nav-button/NavButton.style.tsx
@@ -1,7 +1,7 @@
+import { mqDown } from '@utils';
+
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
-
-import { mqDown } from '@utils/index';
 
 export const NavButton = styled.button`
   ${({ theme }) => css`
