+++ frontend/src/layout/main/Main.style.tsx
@@ -1,6 +1,6 @@
-import styled from '@emotion/styled';
+import { mqDown } from '@utils';
 
-import { mqDown } from '@utils/index';
+import styled from '@emotion/styled';
 
 export const Main = styled.main`
   padding: 120px 0 80px;
