+++ frontend/src/mocks/reviewHandler.ts
@@ -1,8 +1,7 @@
+import { isObject } from '@utils';
 import { rest } from 'msw';
 
 import reviewJSON from '@mocks/reviews.json';
-
-import isObject from '@utils/isObject';
 
 export const reviewHandlers = [
   rest.get('/api/studies/:studyId/reviews', (req, res, ctx) => {
