+++ frontend/src/pages/create-study-page/CreateStudyPage.style.tsx
@@ -1,7 +1,7 @@
+import { mqDown } from '@utils';
+
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
-
-import { mqDown } from '@utils/index';
 
 import { DescriptionTab } from '@create-study-page/components/description-tab/DescriptionTab.style';
