+++ frontend/src/pages/create-study-page/components/period/Peroid.tsx
@@ -1,8 +1,7 @@
+import { getNextYear, getToday } from '@utils';
 import { useMemo } from 'react';
 
 import { css } from '@emotion/react';
-
-import { getNextYear, getToday } from '@utils/index';
 
 import { useFormContext } from '@hooks/useForm';
