+++ frontend/src/pages/create-study-page/hooks/useFetchTagList.ts
@@ -2,7 +2,7 @@ import { useQuery } from 'react-query';
 
 import type { GetTagListResponseData } from '@custom-types';
 
-import { getTagList } from '@api/getTagList';
+import { getTagList } from '@api';
 
 const useFetchTagList = () => {
   return useQuery<GetTagListResponseData, Error>('filters', getTagList);
