+++ frontend/src/pages/create-study-page/hooks/usePostNewStudy.ts
@@ -2,7 +2,7 @@ import { useMutation } from 'react-query';
 
 import type { EmptyObject, PostNewStudyRequestBody } from '@custom-types';
 
-import postNewStudy from '@api/postNewStudy';
+import { postNewStudy } from '@api';
 
 const usePostNewStudy = () => {
   return useMutation<EmptyObject, Error, PostNewStudyRequestBody>(postNewStudy);
