+++ frontend/src/pages/detail-page/DetailPage.style.tsx
@@ -1,7 +1,7 @@
+import { mqDown } from '@utils';
+
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
-
-import { mqDown } from '@utils/index';
 
 export const Main = styled.div`
   display: flex;
