+++ frontend/src/pages/detail-page/components/study-member-section/StudyMemberSection.style.tsx
@@ -1,7 +1,7 @@
+import { mqDown } from '@utils';
+
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
-
-import { mqDown } from '@utils/index';
 
 export const StudyMemberSection = styled.section`
   padding: 16px;
