+++ frontend/src/pages/detail-page/hooks/useFetchStudyReviews.ts
@@ -4,7 +4,7 @@ import { QK_FETCH_STUDY_REVIEWS } from '@constants';
 
 import type { GetReviewResponseData } from '@custom-types';
 
-import getStudyReviews from '@api/getStudyReviews';
+import { getStudyReviews } from '@api';
 
 const useFetchStudyReviews = (studyId: number, size?: number) => {
   const queryKey = size ? [QK_FETCH_STUDY_REVIEWS, studyId] : [QK_FETCH_STUDY_REVIEWS, studyId, 'all'];
