+++ frontend/src/pages/index.tsx
@@ -1,14 +1,7 @@
-import ErrorPage from '@pages/error-page/ErrorPage';
-import LoginRedirectPage from '@pages/login-redirect-page/LoginRedirectPage';
-
-import MainPage from '@main-page/MainPage';
-
-import DetailPage from '@detail-page/DetailPage';
-
-import CreateStudyPage from '@create-study-page/CreateStudyPage';
-
-import MyStudyPage from '@my-study-page/MyStudyPage';
-
-import StudyRoomPage from '@study-room-page/StudyRoomPage';
-
-export { CreateStudyPage, DetailPage, ErrorPage, LoginRedirectPage, MainPage, MyStudyPage, StudyRoomPage };
+export { default as ErrorPage } from '@pages/error-page/ErrorPage';
+export { default as LoginRedirectPage } from '@pages/login-redirect-page/LoginRedirectPage';
+export { default as MainPage } from '@main-page/MainPage';
+export { default as DetailPage } from '@detail-page/DetailPage';
+export { default as CreateStudyPage } from '@create-study-page/CreateStudyPage';
+export { default as MyStudyPage } from '@my-study-page/MyStudyPage';
+export { default as StudyRoomPage } from '@study-room-page/StudyRoomPage';
