+++ frontend/src/pages/main-page/MainPage.style.tsx
@@ -1,6 +1,6 @@
-import styled from '@emotion/styled';
+import { mqDown } from '@utils';
 
-import { mqDown } from '@utils/index';
+import styled from '@emotion/styled';
 
 export const CardList = styled.ul`
   display: grid;
