+++ frontend/src/pages/main-page/components/filter-section/FilterSection.style.tsx
@@ -1,7 +1,7 @@
+import { mqDown } from '@utils';
+
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
-
-import { mqDown } from '@utils/index';
 
 export const FilterSectionContainer = styled.div`
   ${({ theme }) => css`
