+++ frontend/src/pages/main-page/components/filter-section/FilterSection.tsx
@@ -1,9 +1,8 @@
+import { getTagList } from '@api';
 import { useRef } from 'react';
 import { useQuery } from 'react-query';
 
 import type { GetTagListResponseData, Tag, TagInfo } from '@custom-types';
-
-import { getTagList } from '@api/getTagList';
 
 import ArrowButton from '@components/arrow-button/ArrowButton';
