+++ frontend/src/pages/main-page/components/filter-section/filter-button-list/FilterButtonList.style.tsx
@@ -1,6 +1,6 @@
-import styled from '@emotion/styled';
+import { mqDown } from '@utils';
 
-import { mqDown } from '@utils/index';
+import styled from '@emotion/styled';
 
 export const FilterButtons = styled.ul`
   display: flex;
