+++ frontend/src/pages/my-study-page/MyStudyPage.tsx
@@ -4,7 +4,7 @@ import { css } from '@emotion/react';
 
 import type { GetMyStudyResponseData, MyStudy } from '@custom-types';
 
-import getMyStudyList from '@api/getMyStudyList';
+import { getMyStudyList } from '@api';
 
 import Divider from '@components/divider/Divider';
 import Wrapper from '@components/wrapper/Wrapper';
