+++ frontend/src/utils/index.ts
@@ -1,5 +1,4 @@
-import { changeDateSeperator, getNextYear, getToday, yyyymmddTommdd } from '@utils/dates';
-import { mqBetween, mqDown, mqUp } from '@utils/media-query';
-import noop from '@utils/noop';
-
-export { mqDown, mqUp, mqBetween, noop, yyyymmddTommdd, changeDateSeperator, getNextYear, getToday };
+export { default as noop } from '@utils/noop';
+export { mqBetween, mqDown, mqUp } from '@utils/media-query';
+export { changeDateSeperator, getNextYear, getToday, yyyymmddTommdd } from '@utils/dates';
+export { isObject, isString } from '@utils/type-checks';
