--- frontend/src/utils/type-check.ts
@@ -1,1 +1,0 @@
-export const isString = (value: unknown): value is string => typeof value === 'string';
