+++ frontend/tsconfig.json
@@ -21,8 +21,10 @@
       "@styles/*": ["styles/*"],
       "@custom-types": ["custom-types/index.ts"],
       "@assets/*": ["assets/*"],
+      "@utils": ["utils/index.ts"],
       "@utils/*": ["utils/*"],
       "@constants": ["constants.ts"],
+      "@api": ["api/index.ts"],
       "@api/*": ["api/*"],
       "@context/*": ["context/*"],
       "@main-page/*": ["pages/main-page/*"],
